requirements.txt
accelerate==1.3.0
antlr4-python3-runtime==4.9.3
bitsandbytes==0.45.1
clean-fid>=0.1.35
clip-score @ git+https://github.com/AniAggarwal/clip-score.git
contourpy==1.3.1
cycler==0.12.1
diffusers==0.32.2
huggingface_hub==0.28.0
img2dataset>=1.47.0
lightning-utilities==0.14.3
matplotlib==3.10.1
natsort==8.4.0
numpy==2.0.1
omegaconf==2.3.0
open_clip_torch==2.32.0
opencv-python==4.11.0.86
pandas==2.2.3
peft>=0.17.1
pillow==11.1.0
pycocotools==2.0.8
pyparsing==3.2.1
safetensors==0.4.5
scikit-image==0.25.1
scipy==1.15.1
seaborn==0.13.2
sentencepiece==0.2.0
sympy==1.13.1
timm>=1.0.24
tokenizers==0.21.0
torch==2.5.1
torch-fidelity==0.3.0
torcheval==0.0.7
torchmetrics==1.7.1
torchvision==0.20.1
transformers==4.48.2
triton==3.1.0
